Added MealPlanner.java
This is a fragment used to display the main UI for the meal planner part of the app. It contains 5 buttons:

Current Meal Plan - Displays an interactable list of the current meals.
Add - Adds a meal to the list.
Delete - Deletes a meal from the list.
Clear - Deletes all meals from the list.
Back To Home - Returns user to previous screen.