Added MapsFragment.java
This is the fragment for displaying Google Maps. Upon opening, it checks if you have location permissions enabled. If not, it prompts you to enable them. Once enabled, it pans to your location and places a marker on every place designated as a "gym" within a 5km radius, and says whether the place is open or not (when the marker is tapped).